final commit

+++ backend/src/api.py
@@ -17,7 +17,7 @@ CORS(app)
 !! NOTE THIS MUST BE UNCOMMENTED ON FIRST RUN
 !! Running this funciton will add one
 '''
-#db_drop_and_create_all()
+db_drop_and_create_all()
 
 # ROUTES
 '''
@@ -31,20 +31,17 @@ CORS(app)
 @app.route('/drinks', methods=['GET'])
 def drinks():
 
-
-
     drinks = Drink.query.all()
     drink_short = []
     for i in drinks:
         drink_short.append(i.short())
 
 
-
     return jsonify({
         'success': True,
         'status': 200,
         'drinks': drink_short
-    })
+    }), 200
 
 
 
@@ -52,7 +49,6 @@ def drinks():
 @app.route('/drinks-detail')
 @requires_auth('get:drinks-detail')
 def drinks_detail(payload):
-
 
 
     drinks = Drink.query.all()
@@ -66,36 +62,20 @@ def drinks_detail(payload):
         'success': True,
         'status': 200,
         'drinks': drink_long
-    })
-
-'''
-@TODO implement endpoint
-    GET /drinks-detail
-        it should require the 'get:drinks-detail' permission
-        it should contain the drink.long() data representation
-    returns status code 200 and json {"success": True, "drinks": drinks} where drinks is the list of drinks
-        or appropriate status code indicating reason for failure
-'''
+    }), 200
 
 
-'''
-@TODO implement endpoint
-    POST /drinks
-        it should create a new row in the drinks table
-        it should require the 'post:drinks' permission
-        it should contain the drink.long() data representation
-    returns status code 200 and json {"success": True, "drinks": drink} where drink an array containing only the newly created drink
-        or appropriate status code indicating reason for failure
-'''
+
+
 @app.route('/drinks', methods=['POST'])
 @requires_auth('post:drinks')
 def post_drink(payload):
 
-    body = request.get_json()
-
     try:
+        body = request.get_json()
         new_title = body.get('title', None)
         new_recipe = body.get('recipe', None)
+
 
         new_drink = Drink(title=new_title, recipe=json.dumps(new_recipe))
         new_drink.insert()
@@ -107,23 +87,13 @@ def post_drink(payload):
             'success': True,
             'status': 200,
             'drinks': drink_long
-        })
+        }), 200
 
     except:
         abort(422)
-'''
-@TODO implement endpoint
-    PATCH /drinks/<id>
-        where <id> is the existing model id
-        it should respond with a 404 error if <id> is not found
-        it should update the corresponding row for <id>
-        it should require the 'patch:drinks' permission
-        it should contain the drink.long() data representation
-    returns status code 200 and json {"success": True, "drinks": drink} where drink an array containing only the updated drink
-        or appropriate status code indicating reason for failure
-'''
 
-@app.route('/drinks/<drink_id>')
+
+@app.route('/drinks/<drink_id>', methods=['PATCH'])
 @requires_auth('patch:drinks')
 def edit_drink(payload, drink_id):
 
@@ -153,22 +123,14 @@ def edit_drink(payload, drink_id):
 
 
 
-'''
-@TODO implement endpoint
-    DELETE /drinks/<id>
-        where <id> is the existing model id
-        it should respond with a 404 error if <id> is not found
-        it should delete the corresponding row for <id>
-        it should require the 'delete:drinks' permission
-    returns status code 200 and json {"success": True, "delete": id} where id is the id of the deleted record
-        or appropriate status code indicating reason for failure
-'''
 
 @app.route('/drinks/<drink_id>', methods=['DELETE'])
 @requires_auth('delete:drinks')
 def delete_drink(payload, drink_id):
 
     drink_to_delete = Drink.query.filter(Drink.id == drink_id).one_or_none()
+    if drink_to_delete is None:
+        abort(404)
 
     try:
         drink_to_delete.delete()
@@ -177,16 +139,14 @@ def delete_drink(payload, drink_id):
             'success':True,
             'status': 200,
             'delete': drink_id
-        })
+        }), 200
     except:
         abort(422)
 
 
 
 # Error Handling
-'''
-Example error handling for unprocessable entity
-'''
+
 
 
 @app.errorhandler(422)
@@ -206,17 +166,11 @@ def not_found(error):
         "message": "resource not found"
     }), 404
 
-@app.errorhandler(422)
-def unprocessable(error):
-    return jsonify({
-      "success": False,
-      "error": 422,
-      "message": "unprocessable"
-    }), 422
 
 
 @app.errorhandler(405)
-def not_found(error):
+def method_not_allowed(error):
+
     return jsonify({
       "success": False,
       "error": 405,
@@ -227,33 +181,14 @@ def not_found(error):
 @app.errorhandler(AuthError)
 def authentification_error(error):
 
-
-    return jsonify({
-      "success": False,
-      "error": error.status_code,
-      "message": error.error,
-    }), 401
+    status_code = error.status_code
+    msg = jsonify(error.error)
+    return msg, status_code
 
 
 
-'''
-@TODO implement error handlers using the @app.errorhandler(error) decorator
-    each error handler should return (with approprate messages):
-             jsonify({
-                    "success": False,
-                    "error": 404,
-                    "message": "resource not found"
-                    }), 404
-
-'''
-
-'''
-@TODO implement error handler for 404
-    error handler should conform to general task above
-'''
 
 
-'''
-@TODO implement error handler for AuthError
-    error handler should conform to general task above
-'''
+
+
+
